+++ pkgs/fake_async/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 1.3.2-wip
 
 * Require Dart 3.3
+* Fix bug where a `flushTimers` or `elapse` call from within
+  the callback of a periodic timer would immediately invoke
+  the same timer.
 
 ## 1.3.1